requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
config==0.5.1
decorator==5.1.1
dm-tree==0.1.8
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
mido==1.3.0
ml-dtypes==0.2.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pip==23.3.1
pretty-midi==0.2.10
progress==1.5
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.0.0
six==1.16.0
tensorflow-probability==0.23.0
termcolor==2.4.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.14.1

